Update python code for schema rework
This commit updates the python implementation of the LOIN schema to the
latest version, where the correspondence to the pdt schema is in the
SpecificationPerObjectType. The code has been changed to reflect that
the norm is now ISO 7817-3.
The only deviation from the xsd is that the AlphanumericalInformation
type has been renamed to InformationContent to distinguish from the old
terminology. This can be adjusted!
Also the 20240915_instance.py file uses the new implementation to
reproduce the ISO-7817-3_instance.xml example. The resulting xml is a
bit more verbose, but this is due to me not looking into the namespace
functionality of the `xsdata` module, but else they are matching.
The code to reproduce the example is probably hard to read for
outsiders, but as a proof of concept this is sufficient.